requirements.txt
fastapi
uvicorn
pydantic
requests
traci
sumolib
gymnasium
numpy
pyyaml
matplotlib
seaborn
pygame-ce